Make authn and authz handling explicit through errors (#149)
The multiple concatenated routes can yield arbitrary rejections
if route processing is not halted at the first authn or authz
rejection.

For example yielding an AuthenticationFailedRejection will not
complete the route if there's another route that matches the same
path but with a different method. The result in this case will be
MethodNotAllowed when actually the expected result should be
Unauthorized.